add labeling skeleton

+++ Notes/ContentView.swift
@@ -10,13 +10,31 @@ import SwiftData
 
 struct ContentView: View {
 	@Environment(\.modelContext) private var modelContext
-	@Query private var items: [Note]
+	@Environment(\.editMode) private var editMode
+	@Query public var items: [Note]
+	@State private var selection: Set<UUID> = []
+
+	@State private var navigationPath: NavigationPath = NavigationPath()
+
+	var isEditMode: Bool {
+		editMode?.wrappedValue.isEditing ?? false
+	}
 
 	var body: some View {
-		NavigationStack {
-			List {
-				ForEach(items) { note in
-					NavigationLink(note.title, value: note)
+		NavigationStack (path: $navigationPath) {
+			List(selection: $selection) {
+				ForEach(items, id: \.id) { note in
+					HStack {
+						Text(note.title)
+						Spacer()
+					}
+					.contentShape(Rectangle())
+					.tag(note.id)
+					.onTapGesture {
+						if isEditMode { return }
+						navigationPath.append(note)
+					}
+					
 				}
 			}
 			.toolbar {
@@ -35,8 +53,10 @@ struct ContentView: View {
 				}
 				ToolbarItem {
 					Button {
-						for item in items {
-							modelContext.delete(item)
+						for id in selection {
+							if let item = items.first(where: { $0.id == id }) {
+								modelContext.delete(item)
+							}
 						}
 						do {
 							try modelContext.save()
@@ -46,6 +66,9 @@ struct ContentView: View {
 					} label: {
 						Image(systemName: "trash")
 					}
+				}
+				ToolbarItem {
+					EditButton()
 				}
 			}
 			.navigationDestination(for: Note.self) { note in
